requirements.txt
streamlit==1.32.0
pandas==2.1.4
numpy==1.24.3
yfinance==0.2.32
google-generativeai==0.3.0
feedparser==6.0.10
requests==2.31.0
python-dotenv==1.0.0
plotly==5.18.0
